+++ src/phpDebug.ts
@@ -98,6 +98,10 @@ export interface LaunchRequestArguments extends VSCodeDebugProtocol.LaunchReques
     maxConnections?: number
     /** Xdebug cloud token */
     xdebugCloudToken?: string
+    /** Xdebug stream settings */
+    stream?: {
+        stdout?: 0 | 1 | 2
+    }
 
     // CLI options
 
@@ -439,6 +443,11 @@ class PhpDebugSession extends vscode.DebugSession {
                     throw new Error('Cannot have port and socketPath set at the same time')
                 }
                 if (args.hostname?.toLowerCase()?.startsWith('unix://') === true) {
+                    if (fs.existsSync(args.hostname.substring(7))) {
+                        throw new Error(
+                            `File ${args.hostname.substring(7)} exists and cannot be used for Unix Domain socket`
+                        )
+                    }
                     server.listen(args.hostname.substring(7))
                 } else if (args.hostname?.startsWith('\\\\') === true) {
                     server.listen(args.hostname)
@@ -599,6 +608,15 @@ class PhpDebugSession extends vscode.DebugSession {
             this._args.xdebugSettings = xdebugSettings
         } catch (error) {
             throw new Error(`Error applying xdebugSettings: ${String(error instanceof Error ? error.message : error)}`)
+        }
+
+        const stdout =
+            this._args.stream?.stdout === undefined ? (this._args.externalConsole ? 1 : 0) : this._args.stream.stdout
+        if (stdout) {
+            await connection.sendStdout(stdout)
+            connection.on('stream', (stream: xdebug.Stream) =>
+                this.sendEvent(new vscode.OutputEvent(stream.value, 'stdout'))
+            )
         }
 
         this.sendEvent(new vscode.ThreadEvent('started', connection.id))