+++ src/test/adapter.ts
@@ -5,7 +5,8 @@ import { DebugClient } from '@vscode/debugadapter-testsupport'
 import { DebugProtocol } from '@vscode/debugprotocol'
 import * as semver from 'semver'
 import * as net from 'net'
-import { describe, it, beforeEach, afterEach } from 'mocha'
+import * as childProcess from 'child_process'
+import { describe, it, beforeEach, afterEach, after } from 'mocha'
 chai.use(chaiAsPromised)
 const assert = chai.assert
 
@@ -91,6 +92,16 @@ describe('PHP Debug Adapter', () => {
                 client.configurationSequence(),
                 client.waitForEvent('terminated'),
             ])
+        })
+        ;(process.platform === 'win32' ? it.skip : it)('should error on existing unix pipe', async () => {
+            await assert.isRejected(
+                client.launch({
+                    program,
+                    hostname: 'unix:///tmp',
+                    runtimeArgs: ['-dxdebug.client_host=unix:///tmp'],
+                }),
+                /File .+ exists and cannot be used for Unix Domain socket/
+            )
         })
     })
 
@@ -819,6 +830,20 @@ describe('PHP Debug Adapter', () => {
             await Promise.all([client.launch({ program }), client.configurationSequence()])
             await client.assertOutput('stdout', 'stdout output 1\nstdout output 2')
             await client.assertOutput('stderr', 'stderr output 1\nstderr output 2')
+        })
+    })
+
+    describe('stream tests', () => {
+        const program = path.join(TEST_PROJECT, 'output.php')
+
+        it('listen with externalConsole', async () => {
+            // this is how we can currently turn on stdout redirect
+            await Promise.all([client.launch({ stream: { stdout: '1' } }), client.configurationSequence()])
+
+            const script = childProcess.spawn('php', [program])
+            after(() => script.kill())
+            await client.assertOutput('stdout', 'stdout output 1')
+            await client.assertOutput('stdout', 'stdout output 2')
         })
     })