+++ src/xdebugConnection.ts
@@ -182,6 +182,24 @@ export class UserNotify extends Notify {
     }
 }
 
+export class Stream {
+    /** Type of stream */
+    type: string
+    /** Data of the stream */
+    value: string
+
+    /** Constructs a stream object from an XML node from a Xdebug response */
+    constructor(document: XMLDocument) {
+        this.type = document.documentElement.getAttribute('type')!
+        const encoding = document.documentElement.getAttribute('encoding')
+        if (encoding) {
+            this.value = iconv.encode(document.documentElement.textContent!, encoding).toString()
+        } else {
+            this.value = document.documentElement.textContent!
+        }
+    }
+}
+
 export type BreakpointType = 'line' | 'call' | 'return' | 'exception' | 'conditional' | 'watch'
 export type BreakpointState = 'enabled' | 'disabled'
 export type BreakpointResolved = 'resolved' | 'unresolved'
@@ -748,6 +766,7 @@ export declare interface Connection extends DbgpConnection {
     on(event: 'log', listener: (text: string) => void): this
     on(event: 'notify_user', listener: (notify: UserNotify) => void): this
     on(event: 'notify_breakpoint_resolved', listener: (notify: BreakpointResolvedNotify) => void): this
+    on(event: 'stream', listener: (stream: Stream) => void): this
 }
 
 /**
@@ -812,6 +831,9 @@ export class Connection extends DbgpConnection {
             } else if (response.documentElement.nodeName === 'notify') {
                 const n = Notify.fromXml(response, this)
                 this.emit('notify_' + n.name, n)
+            } else if (response.documentElement.nodeName === 'stream') {
+                const s = new Stream(response)
+                this.emit('stream', s)
             } else {
                 const transactionId = parseInt(response.documentElement.getAttribute('transaction_id')!)
                 if (this._pendingCommands.has(transactionId)) {
@@ -826,7 +848,13 @@ export class Connection extends DbgpConnection {
                 }
             }
         })
-        this.on('close', () => this._initPromiseRejectFn(new Error('connection closed (on close)')))
+        this.on('close', () => {
+            this._pendingCommands.forEach(command => command.rejectFn(new Error('connection closed (on close)')))
+            this._pendingCommands.clear()
+            this._commandQueue.forEach(command => command.rejectFn(new Error('connection closed (on close)')))
+            this._commandQueue = []
+            this._initPromiseRejectFn(new Error('connection closed (on close)'))
+        })
     }
 
     /** Returns a promise that gets resolved once the init packet arrives */
@@ -1111,5 +1139,15 @@ export class Connection extends DbgpConnection {
     /** sends an eval command */
     public async sendEvalCommand(expression: string): Promise<EvalResponse> {
         return new EvalResponse(await this._enqueueCommand('eval', undefined, expression), this)
+    }
+
+    // ------------------------------ stream ----------------------------------------
+
+    public async sendStdout(mode: 0 | 1 | 2): Promise<Response> {
+        return new Response(await this._enqueueCommand('stdout', `-c ${mode}`), this)
+    }
+
+    public async sendStderr(mode: 0 | 1 | 2): Promise<Response> {
+        return new Response(await this._enqueueCommand('stderr', `-c ${mode}`), this)
     }
 }